requirements.txt
requests
urllib3==1.26.15
dnacentersdk
pyopenssl
python-dotenv
PyYAML
PyGithub
ciscoconfparse
openai

chromadb
langchain
unstructured
sentence-transformers
langchain-openai
langchain_community
langchain_chroma
langchain_huggingface
ollama
libmagic